added menhir and weapons positions updates
